ai-agents-of-change/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
aiohttp==3.9.3
certifi==2024.2.2
PyPDF2==3.0.1
jinja2==3.1.2
python-multipart==0.0.6 